requirements.txt
httpx>=0.27,<1.0
openai>=1.30,<2.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0
